Lab5 Update

+++ src/main/java/pl/edu/pg/eti/kask/forge/errand/service/ErrandService.java
@@ -5,11 +5,17 @@ import pl.edu.pg.eti.kask.forge.equipment.entity.Equipment;
 import pl.edu.pg.eti.kask.forge.equipment.repository.EquipmentRepository;
 import pl.edu.pg.eti.kask.forge.errand.entity.Errand;
 import pl.edu.pg.eti.kask.forge.errand.repository.ErrandRepository;
+import pl.edu.pg.eti.kask.forge.user.entity.Role;
 import pl.edu.pg.eti.kask.forge.user.entity.User;
 import pl.edu.pg.eti.kask.forge.user.repository.UserRepository;
 
+import javax.annotation.security.RolesAllowed;
+import javax.ejb.EJBAccessException;
+import javax.ejb.LocalBean;
+import javax.ejb.Stateless;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
+import javax.security.enterprise.SecurityContext;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
@@ -17,8 +23,10 @@ import java.util.Optional;
 /**
  * Service layer for all business actions regarding user entity.
  */
-@ApplicationScoped
+@Stateless
+@LocalBean
 @NoArgsConstructor//Empty constructor is required for creating proxy while CDI injection.
+@RolesAllowed(Role.USER)
 public class ErrandService {
 
     /**
@@ -27,39 +35,59 @@ public class ErrandService {
     private ErrandRepository errandRepository;
     private EquipmentRepository equipmentRepository;
     private UserRepository userRepository;
+    /**
+     * Build in security context.
+     */
+    private SecurityContext securityContext;
 
 
     /**
      * @param errandRepository repository for user entity
      */
     @Inject
-    public ErrandService(ErrandRepository errandRepository, EquipmentRepository equipmentRepository, UserRepository userRepository) {
+    public ErrandService(ErrandRepository errandRepository, EquipmentRepository equipmentRepository,
+                         UserRepository userRepository, SecurityContext securityContext) {
         this.errandRepository = errandRepository;
         this.equipmentRepository = equipmentRepository;
         this.userRepository = userRepository;
+        this.securityContext = securityContext;
     }
 
     /**
      * @param id existing username
      * @return container (can be empty) with user
      */
+    //@RolesAllowed(Role.SYSTEM_ADMIN)
     public Optional<Errand> find(Long id) {
-        return errandRepository.find(id);
+        if(securityContext.isCallerInRole(Role.SYSTEM_ADMIN))
+            return errandRepository.find(id);
+        else
+            return errandRepository.findByIdAndUser(id, userRepository.find(securityContext.getCallerPrincipal().getName()).orElseThrow());
+    }
+
+    public Optional<Errand> findForCallerPrincipal(Long id){
+        return errandRepository.findByIdAndUser(id, userRepository.find(securityContext.getCallerPrincipal().getName()).orElseThrow());
     }
 
     /**
      * @return list of users
      */
+    //@RolesAllowed(Role.SYSTEM_ADMIN)
     public List<Errand> findAll() {
-        return errandRepository.findAll();
+        if(securityContext.isCallerInRole(Role.SYSTEM_ADMIN))
+            return errandRepository.findAll();
+        else
+            return errandRepository.findAllByUser(userRepository.find(securityContext.getCallerPrincipal().getName()).orElseThrow());
     }
 
+    public List<Errand> findAllForCallerPrincipal(Long id){
+        return errandRepository.findAllByUser(userRepository.find(securityContext.getCallerPrincipal().getName()).orElseThrow());
+    }
     /**
      * Saves new user.
      *
      * @param errand new user to be saved
      */
-    @Transactional
     public void create(Errand errand) {
         errandRepository.create(errand);
         equipmentRepository.find(errand.getEquipment().getId()).ifPresent(
@@ -70,29 +98,56 @@ public class ErrandService {
         );
     }
 
+    public void createForCallerPrincipal(Errand errand) {
+        User user = userRepository.find(securityContext.getCallerPrincipal().getName()).orElseThrow();
+        errand.setUser(user);
+        user.getErrands().add(errand);
+        errandRepository.create(errand);
+    }
+
     /**
      * Update user.
      *
      * @param errand new user to be saved
      */
-    @Transactional
     public void update(Errand errand) {
         Errand original = errandRepository.find(errand.getId()).orElseThrow();
-        errandRepository.detach(original);
-        errandRepository.update(errand);
+
+        if ((securityContext.isCallerInRole(Role.USER) && original.getUser().getLogin().equals(securityContext.getCallerPrincipal().getName()))
+                || securityContext.isCallerInRole(Role.SYSTEM_ADMIN)) {
+            errandRepository.detach(original);
+            if (!original.getUser().getLogin().equals(errand.getUser().getLogin())) {
+                original.getUser().getErrands().removeIf(errandToRemove -> errandToRemove.getId() == errand.getId());
+                userRepository.find(errand.getUser().getLogin()).ifPresent(user -> user.getErrands().add(errand));
+            }
+            errandRepository.update(errand);
+        } else {
+            throw new EJBAccessException("Authorization failed for user " + securityContext.getCallerPrincipal().getName());
+        }
     }
 
-    /**
-     * Delete user.
-     *
-     * @param id new user to be saved
-     */
-    @Transactional
     public void delete(Long id) {
         Errand errand = errandRepository.find(id).orElseThrow();
-        errand.getEquipment().getErrands().remove(errand);
-        errand.getUser().getErrands().remove(errand);
-        errandRepository.delete(errand);
+        if ((securityContext.isCallerInRole(Role.USER) && errand.getUser().getLogin().equals(securityContext.getCallerPrincipal().getName()))
+                || securityContext.isCallerInRole(Role.SYSTEM_ADMIN)) {
+            errand.getEquipment().getErrands().remove(errand);
+            errand.getUser().getErrands().remove(errand);
+            errandRepository.delete(errand);
+        }
+        else {
+            throw new EJBAccessException("Authorization failed for user " + securityContext.getCallerPrincipal().getName());
+        }
+    }
+
+    @RolesAllowed(Role.SYSTEM_ADMIN)
+    public void deleteAll() {
+        List<Errand> errands = errandRepository.findAll();
+
+        for (Errand errand: errands) {
+            errand.getEquipment().getErrands().remove(errand);
+            errand.getUser().getErrands().remove(errand);
+            errandRepository.delete(errand);
+        }
     }
 
     public Optional<List<Errand>> findAllForEquipment(Long equipmentId) {
@@ -101,7 +156,11 @@ public class ErrandService {
             return Optional.empty();
         }
         else {
-            return Optional.of(errandRepository.findAllByEquipment(equipmentId));
+            if(securityContext.isCallerInRole(Role.SYSTEM_ADMIN))
+                return Optional.of(errandRepository.findAllByEquipment(equipmentId));
+            else
+                return Optional.of(errandRepository.findAllByEquipmentAndUser(equipmentId,
+                        userRepository.find(securityContext.getCallerPrincipal().getName()).orElseThrow()));
         }
     }
 
@@ -125,11 +184,15 @@ public class ErrandService {
             return Optional.empty();
         }
         else {
-            return errandRepository.findByEquipment(errandId, equipment.get());
+            if(securityContext.isCallerInRole(Role.SYSTEM_ADMIN))
+                return errandRepository.findByEquipment(errandId, equipment.get());
+            else
+                return errandRepository.findByEquipment(errandId, equipment.get(),
+                        userRepository.find(securityContext.getCallerPrincipal().getName()).orElseThrow());
         }
     }
 
-    @Transactional
+    @RolesAllowed(Role.SYSTEM_ADMIN)
     public void deleteAllForEquipment(Long equipmentId) {
         Optional<Equipment> equipment = equipmentRepository.find(equipmentId);
         equipment.ifPresent(value -> errandRepository.deleteAllByEquipment(value));
